requirements.txt
langchain = "==0.3.4"
openai = "==1.5.2"
python-dotenv = "==1.0.1"